TRAVIS_BRANCH is the target branch for PRs; use TRAVIS_PULL_REQUEST.
We were using TRAVIS_BRANCH to run the build on just our PR, but it's set to
`master` for our PR since that variable is set to the *target* branch, not the
source branch, in case of a PR. We use TRAVIS_PULL_REQUEST now.

+++ travis.sh
@@ -4,17 +4,15 @@ set -ev
 
 ./liftsh test
 
-#if [ "${TRAVIS_PULL_REQUEST}" = "false" ]; then
+if [ "${TRAVIS_PULL_REQUEST}" = "1729" ]; then
   openssl aes-256-cbc -K $encrypted_a177bbd76133_key -iv $encrypted_a177bbd76133_iv -in .credentials.enc -out ~/.credentials -d
 
-  if [ "${TRAVIS_BRANCH}" = "publishable-by-travis" ]; then
+  if [ "${TRAVIS_BRANCH}" = "master" ]; then
     ./liftsh publish
-  fi
-
-  if [ "${TRAVIS_BRANCH}" = "lift_26" ]; then
+  elif [ "${TRAVIS_BRANCH}" = "lift_26" ]; then
     ./liftsh ++2.10.4 "project lift-framework-pre-111" publish
     ./liftsh ++2.11.1 publish
   fi
 
   rm ~/.credentials
-#fi
+fi
